+++ src/main/java/org/example/service/UserService.java
@@ -1,13 +1,9 @@
 package org.example.service;
 
 import jakarta.persistence.EntityNotFoundException;
-import org.example.dto.note.NoteResponse;
 import org.example.dto.token.TokenRequest;
-import org.example.dto.token.UpdatedToken;
 import org.example.dto.user.User;
-import org.example.dto.user.UserApi;
 import org.example.dto.user.UserNew;
-import org.example.dto.user.UserNotes;
 import org.example.entity.TokenEntity;
 import org.example.entity.UserEntity;
 import org.example.repository.TokenRepository;
@@ -15,6 +11,8 @@ import org.example.repository.UserRepository;
 import org.example.utils.JwtUtil;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDateTime;
 
 import static org.example.utils.BCryptUtil.checkPassword;
 import static org.example.utils.BCryptUtil.hashPassword;
@@ -32,7 +30,7 @@ public class UserService {
   }
 
   @Transactional
-  public UserNew addNewUser(User user) {
+  public UserNew registerUser(User user) {
     if (userRepository.existsByEmail(user.email())) {
       throw new EntityNotFoundException("User with email: " + user.email() + " already exists");
     }
@@ -43,7 +41,7 @@ public class UserService {
             user.email(),
             null));
 
-    var api = tokenRepository.save(new TokenEntity(null, newUser, jwtUtil.generateToken(newUser.getEmail(), 1000*60*5L), jwtUtil.generateToken(newUser.getEmail(), 1000*60*60*7*24L),  0L , null));
+    var api = tokenRepository.save(new TokenEntity(null, newUser, jwtUtil.generateToken(newUser.getEmail(), 1000*60*5L), jwtUtil.generateToken(newUser.getEmail(), 1000*60*60*7*24L),  0L ,  LocalDateTime.now()));
 
     return new UserNew(
             newUser.getId(),
@@ -54,22 +52,7 @@ public class UserService {
     );
   }
 
-  @Transactional(readOnly = true)
-  public UserNotes findUserAndGetNotes(String email) {
-    UserEntity user = userRepository
-            .findByEmail(email)
-            .orElseThrow(() -> new EntityNotFoundException("User not found with email: " + email));
-
-      return new UserNotes(
-              user.getName(),
-              user.getNotes().stream().filter(note -> note.getDeletedAt() == null)
-                      .map(note -> new NoteResponse(
-                      note.getId(),
-                      note.getValue())).toList());
-
-  }
-
-  public TokenRequest login(String email, String password) {
+  public TokenRequest loginUser(String email, String password) {
     UserEntity user = userRepository
             .findByEmail(email)
             .orElseThrow(() -> new EntityNotFoundException("User not found with email: " + email));
@@ -82,6 +65,7 @@ public class UserService {
       String newRefreshToken = jwtUtil.generateToken(tokenEntity.getUser().getEmail(), 1000 * 60 * 60 * 7 * 24L);
       tokenEntity.setToken(newAccessToken);
       tokenEntity.setRefreshToken(newRefreshToken);
+      tokenEntity.setLastUsedAt(LocalDateTime.now());
       tokenRepository.save(tokenEntity);
 
       return new TokenRequest(newAccessToken, newRefreshToken);
